Move site-templates loader file
